Moving some files/tests around to address comments
