+++ reporters/src/main/java/com/nextdoor/bender/monitoring/embedded/metrics/CloudWatchEmbeddedMetricsReporterFactory.java
@@ -13,7 +13,7 @@
  *
  */
 
-package com.nextdoor.bender.monitoring.cw_embedded_metrics;
+package com.nextdoor.bender.monitoring.embedded.metrics;
 
 import com.nextdoor.bender.config.AbstractConfig;
 import com.nextdoor.bender.monitoring.ReporterFactory;
