+++ reporters/src/test/resources/com/nextdoor/bender/monitoring/embedded/metrics/metrics_with_stats_many_dimensions.json
